move tester account creation logic from tester into a separate inner accountClaimer module
